// src/app/modules/importacionConsumos/models/CombustibleTarjeta.ts
export interface CombustibleTarjeta {
    NumeIden: number;
    NumeTarj?: number | null;
    NumeMovi?: number | null;
    NombMovi?: string | null;
    ModeMovi?: string | null;
    PateMovi?: string | null;
    NumePres?: number | null;
    NumePers?: number | null;
    NumeDocu?: number | null;
    CantLitrTanq?: number | null;
    ObseProdPerm?: string | null;
    FechEntr?: Date | null;
    FechDevo?: string | null;
    NumeEsta?: number | null;
    NumePinn?: number | null;
    NumeImpoTopeTran?: number | null;
    NumeImpoTopeMens?: number | null;
    DescTarj?: string | null;
  }